git-demo/pyproject.toml
[tool.poetry]
name = "git-demo"
version = "0.1.0"
description = ""
authors = ["youling66 <youling66@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gitpython = "^3.1.42"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
